Ok we have problems
I'm starting to regret the whole super double jump thing. The only one that seems useful is the straight one (not being homophobic) bc the down diagonal is iffy and all the other ones could just be basic double jumps with the speed boost. Also I want to add: An attack cooldown, make it so attack momentum carries through in the right direction (boosts in the attack direction) and different levels of attack (slow, medium and fast) which have different hitboxes, I also want to add a sprint state, and an attack buffering system, when I'm done them all I'll make the colliders actually work and have a hit state and then I guess a knockback state. And then we move on to enemies! And ledge grabs